content-orchestration/backend/requirements.txt
Flask
Flask-Login
google-auth
google-cloud-storage
firebase-admin
Flask-Cors
python-dotenv
Flask-JWT-Extended
gunicorn 